+++ CHANGES.rst
@@ -5,13 +5,14 @@ Development
 -----------
 - Add sqlalchemy extension.
 - Change the way links are added to resource representation.
-
 - Fix issue #6: HTTP 500 error when using POST verb on Item resources.
 - Breaking changes:
 
   - `royal.resource:Base` constructor now takes request as third positional
     argument.
   - `royal.resource:Base.resource_url` and `Base.url` method signatures change.
+
+- Add `presources` script to print all resources declared with royal.
 
 0.7.4
 -----